backend/requirements.txt
fastapi
uvicorn
langchain-groq
tavily-python
python-dotenv
pydantic-settings
sqlalchemy
psycopg2-binary
pydantic
